+++ src/algolib/graphs/TreeGraph.java
@@ -16,7 +16,7 @@ public class TreeGraph<V, E>
     }
 
     @Override
-    public long getVerticesCount()
+    public int getVerticesCount()
     {
         return graph.getVerticesCount();
     }
@@ -28,7 +28,7 @@ public class TreeGraph<V, E>
     }
 
     @Override
-    public long getEdgesCount()
+    public int getEdgesCount()
     {
         return graph.getEdgesCount();
     }
@@ -60,13 +60,13 @@ public class TreeGraph<V, E>
     }
 
     @Override
-    public long getOutputDegree(Vertex<V> vertex)
+    public int getOutputDegree(Vertex<V> vertex)
     {
         return graph.getOutputDegree(vertex);
     }
 
     @Override
-    public long getInputDegree(Vertex<V> vertex)
+    public int getInputDegree(Vertex<V> vertex)
     {
         return graph.getInputDegree(vertex);
     }
